requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
PyMuPDF
boto3
python-dotenv
